pyproject.toml
[project]
name = "s-l-i-c-e"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.9,<4.0"

[[project.authors]]
name = "Felix"
email = "felix@snackbuilds.xyz"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
discord = ">=2.3.2,<3.0.0"
python-dotenv = ">=1.0.1,<2.0.0"
asyncpg = ">=0.30.0,<0.31.0"
tortoise-orm = ">=0.24.0,<0.25.0"
aerich = "^0.8.1"
captcha = ">=0.6.0,<0.7.0"
loguru = ">=0.7.3,<0.8.0"
web3 = ">=7.8.0,<8.0.0"
eth-account = ">=0.13.5,<0.14.0"
hexbytes = ">=1.3.0,<2.0.0"
tomlkit = ">=0.13.2,<0.14.0"
toml = ">=0.10.2,<0.11.0"
openpyxl = "^3.1.5"

[[tool.poetry.packages]]
include = "s"

[tool.poetry.group.dev.dependencies]
bandit = "^1.7.5"
flake8 = "^6.0.0"
black = "^23.1.0"
pytest = "^7.4.0"
coverage = "^7.3.1"
codecov = "^2.1.0"
isort = "^6.0.0"

[tool.aerich]
tortoise_orm = "src.infrastructure.settings.DATABASE"
location = "./migrations"
src_folder = "./."

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
